chore: RestaurantDetailModal TS로 변환
